refactor: simplificar o sistema de envio de packs de stickers para uso exclusivo do método individual
